oren/lab-exams/lab-exam-1/reference/asn3J/requirements.txt
colorzero==2.0
gpiozero==2.0.1
paho-mqtt==2.1.0
pyside2==5.15.2.1
shiboken2==5.15.2.1
setuptools==80.10.2
